fixes on likes logics after refactoring

+++ server/src/controllers/likes/removeLikeOnCommentController.ts
@@ -1,6 +1,12 @@
 import { logger } from "../../utils/helpers";
 import { Like } from "../../database/models";
 import { Response, Request } from "express";
+import {
+  HTTP_400_BAD_REQUEST,
+  HTTP_404_NOT_FOUND,
+  HTTP_200_OK,
+  HTTP_500_INTERNAL_SERVER_ERROR,
+} from "../../utils/api";
 
 /**
  * This sets a like on a comment
@@ -12,16 +18,32 @@ export const removeLikeOnCommentController = async (
   res: Response
 ) => {
   try {
-    const userId = req.cookies.userId;
-    const deletedLike = await Like.findOneAndRemove({
-      userId: userId,
-      commentId: req.params.commentId,
-    });
-    logger.info(`Removed like on comment ${req.params.commentId}`);
+    const profile = req.profile;
+    const { id: commentId } = req.params;
 
-    return res.json(deletedLike);
+    if (!commentId)
+      return res.status(400).json({
+        ...HTTP_400_BAD_REQUEST,
+        details: `Post ID not provided`,
+      });
+
+    const deletedLike = await Like.findOneAndRemove({
+      userId: profile._id,
+      commentId,
+    });
+
+    if (!deletedLike)
+      return res.status(400).json({
+        ...HTTP_404_NOT_FOUND,
+        details: `Like on comment ${commentId} not found`,
+      });
+
+    return res
+      .status(200)
+      .json({ ...HTTP_200_OK, data: { like: deletedLike.getPublicData() } });
   } catch (err) {
-    logger.error(err as string);
-    return res.status(400).send({ status: "Error", details: err });
+    return res
+      .status(500)
+      .json({ ...HTTP_500_INTERNAL_SERVER_ERROR, details: err });
   }
 };
